Dont attempt to delete private messages

+++ src/main/java/com/ampro/weebot/listener/events/BetterMessageEvent.java
@@ -296,7 +296,8 @@ public class BetterMessageEvent extends BetterEvent {
 
     /** Delete the message. */
     public void deleteMessage() {
-        this.message.delete().queue();
+        if (!isPrivate())
+            this.message.delete().queue();
     }
 
     /**
